added auxilary function to handle getting min pass value for process changed all types to long inside of defs and pstat

+++ xv6/include/pstat.h
@@ -5,10 +5,10 @@ struct pstat {
 	int inuse; // whether this slot of the process process table is in use (1 or 0)
 	int pid;   // the PID of each process
 	char name[16];	// name of the process
-	int tickets;	// number of tickets assigned to this process
+	long tickets;	// number of tickets assigned to this process
 	long pass;	// current pass value for the process
 	long stride;	// stride value
-	int n_schedule;	// number of times chosen for scheduling
+	long n_schedule;	// number of times chosen for scheduling
 };
 
 /*
